+++ es_stats/__init__.py
@@ -1,1 +1,4 @@
 from es_stats.version import __version__
+from es_stats.classes import (
+    ClusterHealth, ClusterState, ClusterStats, NodeInfo, NodeStats
+)